drag&drop exercise + delete exercise features completed

+++ frontend/src/components/ExerciseComponent.jsx
@@ -1,10 +1,14 @@
 import { useExercisesContext } from '../hooks/useExercisesContext'
-import { FaTrashAlt, FaEdit } from 'react-icons/fa'
+import { FaTrashAlt, FaEdit, FaGripVertical } from 'react-icons/fa'
 import { useAuthContext } from '../hooks/useAuthContext'
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 import { useNavigate } from 'react-router-dom'
+import { useDrag, useDrop } from 'react-dnd'
 
-const exerciseComponent = ({ exercise, workoutId }) => {
+const itemTypes = {
+  EXERCISE: 'exercise',
+}
+
+const exerciseComponent = ({ exercise, workoutId, index, moveExercise}) => {
   const { _id: exerciseId } = exercise
   const { dispatch } = useExercisesContext()
   const { user } = useAuthContext()
@@ -14,6 +18,13 @@ const exerciseComponent = ({ exercise, workoutId }) => {
     if (!user) {
       return
     }
+
+    const shouldDelete = window.confirm('Are you sure you want to delete this exercise?')
+
+    if (!shouldDelete) {
+      return
+    }
+
     const res = await fetch(`http://localhost:3000/${workoutId}/exercises/${exerciseId}`, {
       method: 'DELETE',
       headers: {
@@ -31,11 +42,36 @@ const exerciseComponent = ({ exercise, workoutId }) => {
     navigate(`/${workoutId}/exercises/${exerciseId}/edit`)
   }
 
+  const [, dragRef] = useDrag({
+    type: itemTypes.EXERCISE,
+    item: { id: exercise._id, index },
+  })
+
+  const [, dropRef] = useDrop({
+    accept: itemTypes.EXERCISE,
+    hover(item, monitor) {
+      const dragIndex = item.index
+      const hoverIndex = index
+
+      if (dragIndex === hoverIndex) {
+        return
+      }
+
+      moveExercise(dragIndex, hoverIndex)
+      item.index = hoverIndex
+    },
+  })
+
+  const ref = (node) => {
+    dragRef(node)
+    dropRef(node)
+  }
+
   return (
-    <div className='exercise-card'>
+    <div className='exercise-card' ref={ref}>
       <div className='exercise-header'>
         <div className='exercise-number'>
-          <h2>1</h2>
+          <h2>{index + 1}</h2>
         </div>
 
         <div className='exercise-title'>
@@ -68,7 +104,11 @@ const exerciseComponent = ({ exercise, workoutId }) => {
       </div>
 
       <div className='exercise-btn-container'>
-        <div></div>
+        
+        <span className='exercise-icon exercise-grip'>
+          <FaGripVertical size={20} />
+        </span>
+
         <button className='primary-button' onClick={() => navigate(`/${workoutId}/exercises/${exerciseId}/show-details`)}>
           Show Details
         </button>
@@ -79,28 +119,6 @@ const exerciseComponent = ({ exercise, workoutId }) => {
       </div>
     </div>
   )
-  // <div className='exercise-container'>
-  //   <div className='exercise-item'>
-  //     <div className='exercise-info'>
-  //       <h4 className='exercise-title'>{exercise.title}</h4>
-  //       <p className='exercise-description'>
-  //         Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quis, asperiores non. Quas molestias sit aliquid facere, a, dignissimos molestiae doloremque
-  //         voluptas ab, placeat eius officiis magnam id consequatur! Quo, quae.*2
-  //       </p>
-  //     </div>
-  //     <div className='exercise-details'>
-  //       <div className='exercise-detail'>
-  //         <strong>Sets:</strong> {exercise.sets}
-  //       </div>
-  //       <div className='exercise-detail'>
-  //         <strong>Reps:</strong> {exercise.reps}
-  //       </div>
-  //       <div className='exercise-detail'>
-  //         <strong>Rest:</strong> {exercise.rest} seconds
-  //       </div>
-  //     </div>
-  //   </div>
-  // </div> 
 }
 export default exerciseComponent
